fix callable field invocation overlap footprints
